Compile fix

+++ server/src/Tcp.cpp
@@ -264,7 +264,7 @@ void TCP::DieWithError(char *errorMessage)
 void TCP::printStats()
 {
     printf("TCP stats:\n");
-    printf("Connections - accepted: %d closed: %d interrupted: %d", accepted, closed, interrupted);
-    printf("Bytes - recv: %d send: %d", recvBytes, sendBytes);
-    printf("Messages - recv: %d send: %d", recvMessages, sendMessages);
+    printf("Connections - accepted: %lld closed: %lld interrupted: %lld", (uint64)accepted, (uint64)closed, (uint64)interrupted);
+    printf("Bytes - recv: %lld send: %lld", (uint64)recvBytes, (uint64)sendBytes);
+    printf("Messages - recv: %lld send: %lld", (uint64)recvMessages, (uint64)sendMessages);
 }
